Allow logging in during SSO promotion flow
