+++ pkg/auth/dependency/interaction/flows/webapp_promote.go
@@ -2,7 +2,9 @@ package flows
 
 import (
 	"github.com/skygeario/skygear-server/pkg/auth/dependency/identity"
+	"github.com/skygeario/skygear-server/pkg/auth/dependency/identity/oauth"
 	"github.com/skygeario/skygear-server/pkg/auth/dependency/interaction"
+	"github.com/skygeario/skygear-server/pkg/auth/dependency/sso"
 	"github.com/skygeario/skygear-server/pkg/auth/event"
 	"github.com/skygeario/skygear-server/pkg/auth/model"
 	"github.com/skygeario/skygear-server/pkg/core/authn"
@@ -71,6 +73,61 @@ func (f *WebAppFlow) PromoteWithLoginID(loginIDKey, loginID string, userID strin
 		Step:  step,
 		Token: token,
 	}, nil
+}
+
+func (f *WebAppFlow) PromoteWithOAuthProvider(userID string, oauthAuthInfo sso.AuthInfo) (*WebAppResult, error) {
+	providerID := oauth.NewProviderID(oauthAuthInfo.ProviderConfig)
+	iden := identity.Spec{
+		Type: authn.IdentityTypeOAuth,
+		Claims: map[string]interface{}{
+			identity.IdentityClaimOAuthProviderKeys: providerID.ClaimsValue(),
+			identity.IdentityClaimOAuthSubjectID:    oauthAuthInfo.ProviderUserInfo.ID,
+			identity.IdentityClaimOAuthProfile:      oauthAuthInfo.ProviderRawProfile,
+			identity.IdentityClaimOAuthClaims:       oauthAuthInfo.ProviderUserInfo.ClaimsValue(),
+		},
+	}
+	var err error
+
+	var i *interaction.Interaction
+	if f.ConflictConfig.Promotion == config.PromotionConflictBehaviorLogin {
+		_, _, err = f.Identities.GetByClaims(authn.IdentityTypeOAuth, iden.Claims)
+		if errors.Is(err, identity.ErrIdentityNotFound) {
+			i, err = f.Interactions.NewInteractionAddIdentity(&interaction.IntentAddIdentity{
+				Identity: iden,
+			}, "", userID)
+		} else if err != nil {
+			return nil, err
+		} else {
+			i, err = f.Interactions.NewInteractionLogin(&interaction.IntentLogin{
+				Identity: iden,
+			}, "")
+		}
+	} else {
+		i, err = f.Interactions.NewInteractionAddIdentity(&interaction.IntentAddIdentity{
+			Identity: iden,
+		}, "", userID)
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	s, err := f.Interactions.GetInteractionState(i)
+	if err != nil {
+		return nil, err
+	} else if s.CurrentStep().Step != interaction.StepCommit {
+		// authenticator is not needed for oauth identity
+		// so the current step must be commit
+		panic("interaction_flow_webapp: unexpected interaction step")
+	}
+
+	i.Extra[WebAppExtraStateAnonymousUserPromotion] = userID
+
+	result, err := f.Interactions.Commit(i)
+	if err != nil {
+		return nil, err
+	}
+
+	return f.afterAnonymousUserPromotion(i, result)
 }
 
 func (f *WebAppFlow) afterAnonymousUserPromotion(i *interaction.Interaction, ir *interaction.Result) (*WebAppResult, error) {
