+++ pkg/auth/dependency/interaction/flows/webapp_sso.go
@@ -152,40 +152,3 @@ func (f *WebAppFlow) UnlinkWithOAuthProvider(userID string, providerConfig confi
 
 	return
 }
-
-func (f *WebAppFlow) PromoteWithOAuthProvider(userID string, oauthAuthInfo sso.AuthInfo) (*WebAppResult, error) {
-	providerID := oauth.NewProviderID(oauthAuthInfo.ProviderConfig)
-	claims := map[string]interface{}{
-		identity.IdentityClaimOAuthProviderKeys: providerID.ClaimsValue(),
-		identity.IdentityClaimOAuthSubjectID:    oauthAuthInfo.ProviderUserInfo.ID,
-		identity.IdentityClaimOAuthProfile:      oauthAuthInfo.ProviderRawProfile,
-		identity.IdentityClaimOAuthClaims:       oauthAuthInfo.ProviderUserInfo.ClaimsValue(),
-	}
-
-	clientID := ""
-	i, err := f.Interactions.NewInteractionAddIdentity(&interaction.IntentAddIdentity{
-		Identity: identity.Spec{
-			Type:   authn.IdentityTypeOAuth,
-			Claims: claims,
-		},
-	}, clientID, userID)
-	if err != nil {
-		return nil, err
-	}
-
-	s, err := f.Interactions.GetInteractionState(i)
-	if err != nil {
-		return nil, err
-	} else if s.CurrentStep().Step != interaction.StepCommit {
-		// authenticator is not needed for oauth identity
-		// so the current step must be commit
-		panic("interaction_flow_webapp: unexpected interaction step")
-	}
-
-	result, err := f.Interactions.Commit(i)
-	if err != nil {
-		return nil, err
-	}
-
-	return f.afterAnonymousUserPromotion(i, result)
-}
